// src/models/Nationality.ts
export type Nationality = 
    "Italian"
  | "French"
  | "German"
  | "Spanish"
  | "Portuguese"
  | "British"
  | "American"
  | "Canadian"
  | "Mexican"
  | "Brazilian"
  | "Argentinian"
  | "Chinese"
  | "Japanese"
  | "Korean"
  | "Indian"
  | "Russian"
  | "Australian"
  | "Egyptian"
  | "Turkish"
  | "South African"